refactor: Moved RelatedRequests
Moved related requests component so its no longer exclusive to party sections
