+++ src/components/PartySections/index.js
@@ -1,2 +1,1 @@
-export { default as PartyInfo } from './PartyInfo';
-export { default as RelatedRequests } from './RelatedRequests';
+export { default } from './PartyInfo';
